polish unit tests
